hide footer when comparison disabled
