+++ client/src/containers/analysis-visualization/analysis-table/footer/index.tsx
@@ -15,7 +15,7 @@ const AnalysisTableFooter: FC = () => {
   });
 
   return (
-    <footer className="flex items-end justify-between text-xs text-gray-500">
+    <footer className="flex justify-between text-xs text-gray-500">
       <ul className="space-y-1">
         {Boolean(scenarioName) && (
           <li className="flex items-center space-x-2">
@@ -24,7 +24,7 @@ const AnalysisTableFooter: FC = () => {
             </span>
             <Separator className="h-[1px] w-[26px] text-gray-200" />
             <span className="text-gray-900">140</span>
-            <Badge className="rounded-sm border border-gray-400 bg-transparent p-1 leading-3 text-gray-400">
+            <Badge className="rounded-sm border border-gray-400 bg-transparent p-1 leading-3 text-gray-400 hover:bg-transparent">
               +70
             </Badge>
             <Separator className="h-[1px] w-[26px] text-gray-200" />
@@ -39,24 +39,6 @@ const AnalysisTableFooter: FC = () => {
           <span className="text-gray-400 line-through">70</span>
         </li>
       </ul>
-
-      <div className="flex items-center space-x-2">
-        <svg
-          xmlns="http://www.w3.org/2000/svg"
-          width="27"
-          height="2"
-          viewBox="0 0 27 2"
-          fill="none"
-        >
-          <path
-            d="M27 1L0.733333 1H0"
-            stroke="#60626A"
-            stroke-dasharray="3.6 2.4"
-            strokeWidth="1.2"
-          />
-        </svg>
-        <span>The years at the right of the doted line are projected</span>
-      </div>
     </footer>
   );
 };
